requirements.txt
torch>=2.1.0
transformers>=4.40.0
diffusers>=0.27.0
accelerate>=0.28.0
datasets>=2.18.0
openai>=1.12.0
Pillow>=10.2.0
pyyaml>=6.0.1
tqdm>=4.66.0
numpy>=1.26.0
scipy>=1.12.0
pandas>=2.2.0
loguru>=0.7.2
